+++ CHANGES.md
@@ -1,6 +1,7 @@
 
 - Add `Uuidm.v7`, create time and random based V7 UUIDs using client provided
   random bytes and time. Thanks to Robin Newton for the patch (#14).
+- Add `Uuidm.v8` to create V8 custom UUIDs.
 - Add `Uuidm.max` the RFC 9569 Max UUID.
 - Add `Uuidm.{variant,version,time_ms}` UUID property accessors.
 - Change `Uuidm.v4_gen` generation strategy.